requirements.txt
annotated-types==0.7.0
anyio==4.6.0
Authlib==1.3.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
grpcio-health-checking==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.10
jiter==0.5.0
openai==1.47.0
pinecone==5.3.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
protobuf==5.28.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
validators==0.34.0
